requirements.txt
flask
gunicorn
pyromod==1.5
tgcrypto
aiohttp
aiofiles
requests
hachoir
cloudscraper
pycryptodome
pytz
motor
pymongo>=4.0
pyrofork
PyJWT
aiogram
beautifulsoup4
httpx
python-telegram-bot==20.6
tqdm
colorama
termcolor
python-dotenv
